Make qdoc generate usable docs for QVulkanInstance
It is not just that the unknown native types (e.g. VkInstance) lead to
showing the type as 'int' in some functions, the bigger problem was that
a number of functions were not present in the generated documentation at
all.

Change-Id: Id71fe20c3a70a8d3b75bbc693a2a7ee94bb74642
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit b62f32a4dd87010e4c51d33ef440e2fbba531361)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>