Skip LGTM analysis for the bootstrap library and tools
This is because the same code is built for the bootstrap library with
non-standard options like disabled threads, which causes false alerts.
So we build the bootstrap tools in the "configure" step.

Additionally we update the configure flags to only the ones that
matter for the CMake build, after the latest configure script updates.

Change-Id: Id4f90db2a16904f6665f3994f23a9f2a974aefc6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>